MemberConfigurator was added
