+++ ObjectPrinting/IBasicConfigurator.cs
@@ -8,5 +8,5 @@ public interface IBasicConfigurator<TOwner>
     IBasicConfigurator<TOwner> Exclude<T>();
     IBasicConfigurator<TOwner> Exclude<T>(Expression<Func<TOwner, T>> expression);
     IMemberConfigurator<TOwner, T> ConfigureProperty<T>(Expression<Func<TOwner, T>> expression);
-    string PrintToString(TOwner obj);
+    PrintingConfig<TOwner> ConfigurePrinter();
 }
